+++ maiconverter/converter/simaitomaima2.py
@@ -1,6 +1,12 @@
-from typing import List
+from typing import List, Tuple, Optional
 
-from ..maima2 import MaiMa2
+from ..maima2 import (
+    MaiMa2,
+    BPM,
+    HoldNote as Ma2HoldNote,
+    TouchHoldNote as Ma2TouchHoldNote,
+    SlideNote as Ma2SlideNote,
+)
 from ..simai import (
     SimaiChart,
     pattern_to_int,
@@ -23,6 +29,10 @@ def simai_to_ma2(simai: SimaiChart, fes_mode: bool = False) -> MaiMa2:
 
     ma2.set_meter(0.0, 4, 4)
     convert_notes(ma2, simai.notes)
+
+    if len(ma2.bpms) != 1:
+        fix_durations(ma2)
+
     return ma2
 
 
@@ -81,3 +91,55 @@ def convert_notes(ma2: MaiMa2, simai_notes: List[SimaiNote]) -> None:
             )
         else:
             print(f"Warning: Unknown note type {note_type}")
+
+
+def fix_durations(ma2: MaiMa2):
+    """Simai note durations (slide delay, slide duration, hold note duration)
+    disregards bpm changes midway, unlike ma2. So we'll have to compensate for those.
+    """
+
+    def bpm_changes(start: float, duration: float) -> List[BPM]:
+        result: List[BPM] = []
+        for bpm in ma2.bpms:
+            if start < bpm.measure < start + duration:
+                result.append(bpm)
+
+        return result
+
+    def compensate_duration(
+        start: float, duration: float, base_bpm: float, changes: List[BPM]
+    ) -> float:
+        new_duration = 0
+
+        note_start = start
+        for bpm in changes:
+            new_duration += (
+                ma2.get_bpm(bpm.measure - 0.0001)
+                * (bpm.measure - note_start)
+                / base_bpm
+            )
+
+            note_start = bpm.measure
+
+        if note_start < start + duration:
+            new_duration += (
+                ma2.get_bpm(note_start + 0.0001)
+                * (start + duration - note_start)
+                / base_bpm
+            )
+
+        return new_duration
+
+    for note in ma2.notes:
+        if isinstance(note, (Ma2HoldNote, Ma2TouchHoldNote, Ma2SlideNote)):
+            bpms = bpm_changes(note.measure, note.duration)
+            if len(bpms) != 0:
+                note.duration = compensate_duration(
+                    note.measure, note.duration, ma2.get_bpm(note.measure), bpms
+                )
+        if isinstance(note, Ma2SlideNote):
+            bpms = bpm_changes(note.measure, note.delay)
+            if len(bpms) != 0:
+                note.delay = compensate_duration(
+                    note.measure, note.delay, ma2.get_bpm(note.measure), bpms
+                )